Mv node classes to files in a directory
